messages: remove DIRECTORY and add 6.0 support
The former opp_msgc tool allowed to specify the generation of C/C++ from
messages files in the current working directory. As that is no longer
possible, the function has been simplified and now builds the files in
the same directory as where they are.

I recognise that this is a step back for msg4 users. A future path would
be to rather copy the source message files to the output folder and then
build the messages there, thus keeping the source tree entirely clean.