fix: remove incorrect include directory, just add .cc to source
Also (slightly) increase verbosity when generating files.

User is responsible for making sure that include directories are set
correctly to allow for including generated files. By using the
target_include_directory call for each of the directories, files can be
added as if they are in the current working directory, which does not
seem like correct behaviour.